requirements.txt
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.7.0
livekit-plugins-silero>=0.6.0
openai>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
flask>=3.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0
eventlet>=0.35.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
